# backend/app/api/static/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Track, Manage, and Organize your Things.",
        "title": "Homebox API",
        "contact": {
            "name": "Homebox Team",
            "url": "https://discord.homebox.software"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/actions/create-missing-thumbnails": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates thumbnails for items that are missing them",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actions"
                ],
                "summary": "Create Missing Thumbnails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ActionAmountResult"
                        }
                    }
                }
            }
        },
        "/v1/actions/ensure-asset-ids": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Ensures all items in the database have an asset ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actions"
                ],
                "summary": "Ensure Asset IDs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ActionAmountResult"
                        }
                    }
                }
            }
        },
        "/v1/actions/ensure-import-refs": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Ensures all items in the database have an import ref",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actions"
                ],
                "summary": "Ensures Import Refs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ActionAmountResult"
                        }
                    }
                }
            }
        },
        "/v1/actions/set-primary-photos": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Sets the first photo of each item as the primary photo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actions"
                ],
                "summary": "Set Primary Photos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ActionAmountResult"
                        }
                    }
                }
            }
        },
        "/v1/actions/wipe-inventory": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Deletes all items in the inventory",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actions"
                ],
                "summary": "Wipe Inventory",
                "parameters": [
                    {
                        "description": "Wipe options",
                        "name": "options",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/v1.WipeInventoryOptions"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ActionAmountResult"
                        }
                    }
                }
            }
        },
        "/v1/actions/zero-item-time-fields": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Resets all item date fields to the beginning of the day",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actions"
                ],
                "summary": "Zero Out Time Fields",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.ActionAmountResult"
                        }
                    }
                }
            }
        },
        "/v1/assets/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get Item by Asset ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.PaginationResult-repo_EntitySummary"
                        }
                    }
                }
            }
        },
...
        "/v1/entities": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Query All Entities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search string",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "items per page",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "tags Ids",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "parent Ids",
                        "name": "parentIds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.EntityListResult"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Create Entity",
                "parameters": [
                    {
                        "description": "Entity Data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo.EntityCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/repo.EntityOut"
                        }
                    }
                }
            }
        },
        "/v1/entities/export": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Export Entities",
                "responses": {
                    "200": {
                        "description": "text/csv",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/entities/fields": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get All Custom Field Names",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/entities/fields/values": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get All Custom Field Values",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field name",
                        "name": "field",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/entities/import": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Import Entities",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV file to upload",
                        "name": "csv",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/entities/prices/sync-all": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item Pricing"
                ],
                "summary": "Sync All Tracked Items Market Prices for Group",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.SyncPricesResponse"
                        }
                    }
                }
            }
        },
        "/v1/entities/prices/sync-bulk": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Item Pricing"
                ],
                "summary": "Sync Selected Items Market Prices in Bulk",
                "parameters": [
                    {
                        "description": "Entity IDs to sync",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.SyncPricesBulkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.SyncPricesResponse"
                        }
                    }
                }
            }
        },
        "/v1/entities/tree": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Get Locations Tree",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "include items in response tree",
                        "name": "withItems",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repo.TreeItem"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/entities/{id}/duplicate": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Entities"
                ],
                "summary": "Duplicate Entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Duplicate Options",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo.DuplicateOptions"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/repo.EntityOut"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "repo.PaginationResult-repo_EntitySummary": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.EntitySummary"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "v1.ActionAmountResult": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer"
                }
            }
        },
...
        "v1.WipeInventoryOptions": {
            "type": "object",
            "properties": {
                "wipeLocations": {
                    "type": "boolean"
                },
                "wipeMaintenance": {
                    "type": "boolean"
                },
                "wipeTags": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}